pyproject.toml
[project]
name = "fastapi-board"
version = "0.1.0"
description = "FastAPI + SQLAlchemy 학습용 게시판 프로젝트"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=4.0,<5.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

# pytest 설정 (Spring의 src/test 역할)
[tool.pytest.ini_options]
# asyncio 모드: auto = 모든 async 테스트를 자동으로 비동기 처리
asyncio_mode = "auto"
# 테스트 파일 위치
testpaths = ["tests"]
